+++ vite.config.ts
@@ -13,4 +13,7 @@ export default defineConfig({
       },
     },
   },
+  optimizeDeps: {
+    include: ['i18next', 'react-i18next']
+  }
 });